Split TrustedRouter adapter tests
